Update Aprialtag_landing_demo.py
First of all, incorrect use of input will cause blocking in the program, so the input function is temporarily canceled and fixed start and end points are set. Second, standardize the writing of message seq, stamp,  and frame_id. The next step should use input function correctly and try to use transitions to change the code.