[Mini Quadrotor][camera]turn on and calibration
firstly, I used the usb_cam ros bag to turn on the usb_cam_elecom. I changed the image_width, height and video_device in the launch file.Add the camera_info_url, which is the inter parameter of the camera. To do calibration, I used the camera_calibration ros bag to get the inter parameter.